Combine gen_ecmult_*.c into precompute.c
