+++ .gitignore
@@ -3,8 +3,7 @@ bench_ecmult
 bench_internal
 tests
 exhaustive_tests
-gen_ecmult_gen_static_prec_table
-gen_ecmult_static_pre_g
+precompute
 valgrind_ctime_test
 *.exe
 *.so
