+++ src/ecmult_gen_static_prec_table.h
@@ -1,4 +1,4 @@
-/* This file was automatically generated by gen_ecmult_gen_static_prec_table. */
+/* This file was automatically generated by precompute. */
 /* See ecmult_gen_impl.h for details about the contents of this file. */
 #ifndef SECP256K1_ECMULT_GEN_STATIC_PREC_TABLE_H
 #define SECP256K1_ECMULT_GEN_STATIC_PREC_TABLE_H
