+++ src/precompute.c
@@ -1,11 +1,13 @@
-/*****************************************************************************************************
- * Copyright (c) 2013, 2014, 2017, 2021 Pieter Wuille, Andrew Poelstra, Jonas Nick, Russell O'Connor *
- * Distributed under the MIT software license, see the accompanying                                  *
- * file COPYING or https://www.opensource.org/licenses/mit-license.php.                              *
- *****************************************************************************************************/
+/******************************************************************************
+ * Copyright (c) 2013, 2014, 2015, 2017, 2021 Pieter Wuille, Andrew Poelstra, *
+ * Jonas Nick, Russell O'Connor, Thomas Daede, Cory Fields.                   *
+ * Distributed under the MIT software license, see the accompanying           *
+ * file COPYING or https://www.opensource.org/licenses/mit-license.php.       *
+ ******************************************************************************/
 
 #include <inttypes.h>
 #include <stdio.h>
+#include <string.h>
 
 /* Autotools creates libsecp256k1-config.h, of which ECMULT_WINDOW_SIZE is needed.
    ifndef guard so downstream users can define their own if they do not use autotools. */
@@ -19,6 +21,8 @@
 #include "field_impl.h"
 #include "group_impl.h"
 #include "ecmult.h"
+#include "ecmult_gen.h"
+#include "ecmult_gen_prec_impl.h"
 
 static void precompute_ecmult_print_table(FILE *fp, const char *name, int window_g, const secp256k1_gej *gen, int with_conditionals) {
     secp256k1_gej gj;
@@ -87,7 +91,7 @@ static int precompute_ecmult(void) {
         return -1;
     }
 
-    fprintf(fp, "/* This file was automatically generated by gen_ecmult_static_pre_g. */\n");
+    fprintf(fp, "/* This file was automatically generated by precompute. */\n");
     fprintf(fp, "/* This file contains an array secp256k1_pre_g with odd multiples of the base point G and\n");
     fprintf(fp, " * an array secp256k1_pre_g_128 with odd multiples of 2^128*G for accelerating the computation of a*P + b*G.\n");
     fprintf(fp, " */\n");
@@ -129,6 +133,78 @@ static int precompute_ecmult(void) {
     return 0;
 }
 
-int main(void) {
-    return precompute_ecmult();
+static int precompute_ecmult_gen(void) {
+    const char outfile[] = "src/ecmult_gen_static_prec_table.h";
+    FILE* fp;
+    int bits;
+
+    fp = fopen(outfile, "w");
+    if (fp == NULL) {
+        fprintf(stderr, "Could not open %s for writing!\n", outfile);
+        return -1;
+    }
+
+    fprintf(fp, "/* This file was automatically generated by precompute. */\n");
+    fprintf(fp, "/* See ecmult_gen_impl.h for details about the contents of this file. */\n");
+    fprintf(fp, "#ifndef SECP256K1_ECMULT_GEN_STATIC_PREC_TABLE_H\n");
+    fprintf(fp, "#define SECP256K1_ECMULT_GEN_STATIC_PREC_TABLE_H\n");
+
+    fprintf(fp, "#include \"group.h\"\n");
+
+    fprintf(fp, "#define S(a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p) "
+            "SECP256K1_GE_STORAGE_CONST(0x##a##u,0x##b##u,0x##c##u,0x##d##u,0x##e##u,0x##f##u,0x##g##u,"
+            "0x##h##u,0x##i##u,0x##j##u,0x##k##u,0x##l##u,0x##m##u,0x##n##u,0x##o##u,0x##p##u)\n");
+
+    fprintf(fp, "#ifdef EXHAUSTIVE_TEST_ORDER\n");
+    fprintf(fp, "static secp256k1_ge_storage secp256k1_ecmult_gen_prec_table[ECMULT_GEN_PREC_N(ECMULT_GEN_PREC_BITS)][ECMULT_GEN_PREC_G(ECMULT_GEN_PREC_BITS)];\n");
+    fprintf(fp, "#else\n");
+    fprintf(fp, "static const secp256k1_ge_storage secp256k1_ecmult_gen_prec_table[ECMULT_GEN_PREC_N(ECMULT_GEN_PREC_BITS)][ECMULT_GEN_PREC_G(ECMULT_GEN_PREC_BITS)] = {\n");
+
+    for (bits = 2; bits <= 8; bits *= 2) {
+        int g = ECMULT_GEN_PREC_G(bits);
+        int n = ECMULT_GEN_PREC_N(bits);
+        int inner, outer;
+
+        secp256k1_ge_storage* table = checked_malloc(&default_error_callback, n * g * sizeof(secp256k1_ge_storage));
+        secp256k1_ecmult_gen_create_prec_table(table, &secp256k1_ge_const_g, bits);
+
+        fprintf(fp, "#if ECMULT_GEN_PREC_BITS == %d\n", bits);
+        for(outer = 0; outer != n; outer++) {
+            fprintf(fp,"{");
+            for(inner = 0; inner != g; inner++) {
+                fprintf(fp, "S(%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32
+                            ",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32",%"PRIx32")",
+                        SECP256K1_GE_STORAGE_CONST_GET(table[outer * g + inner]));
+                if (inner != g - 1) {
+                    fprintf(fp,",\n");
+                } 
+            }
+            if (outer != n - 1) {
+                fprintf(fp,"},\n");
+            } else {
+                fprintf(fp,"}\n");
+            }
+        }
+        fprintf(fp, "#endif\n");
+        free(table);
+    }
+
+    fprintf(fp, "};\n");
+    fprintf(fp, "#endif /* EXHAUSTIVE_TEST_ORDER */\n");
+    fprintf(fp, "#undef SC\n");
+    fprintf(fp, "#endif /* SECP256K1_ECMULT_GEN_STATIC_PREC_TABLE_H */\n");
+    fclose(fp);
+
+    return 0;
+}
+
+int main(int argc, char** argv) {
+    if (argc == 2 && !strcmp(argv[1], "ecmult")) {
+        return precompute_ecmult();
+    } else if (argc == 2 && !strcmp(argv[1], "ecmult_gen")) {
+        return precompute_ecmult_gen();
+    } else {
+        fprintf(stderr, "Usage: ./precompute [ecmult|ecmult_gen]\n");
+        return 1;
+    }
 }
